requirements.txt
altgraph==0.17.4
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
packaging==25.0
pefile==2023.2.7
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.9.0
urllib3==2.5.0
